Get everything compiling and running
