+++ src/token.hpp
@@ -74,8 +74,6 @@ enum class token_type
     uint64,
 };
 
-auto to_string(token_type tt) -> std::string_view;
-
 struct token
 {
     std::string_view text;
@@ -116,52 +114,8 @@ struct token
     }
 };
 
-auto print_tokens(const std::vector<anzu::token>& tokens) -> void;
-
-class tokenstream
-{
-    std::vector<token>::const_iterator d_begin;
-    std::vector<token>::const_iterator d_curr;
-    std::vector<token>::const_iterator d_end;
-
-public:
-    tokenstream(const std::vector<token>& tokens);
-
-    auto valid() const -> bool { return d_curr != d_end; }
-    auto has_next() const -> bool { return valid() && std::next(d_curr) != d_end; }
-
-    auto curr() const -> const token& { return *d_curr; }
-    auto next() const -> const token& { return *std::next(d_curr); }
-    auto position() const -> std::int64_t { return std::distance(d_begin, d_curr) + 1; }
-
-    auto consume() -> token
-    {
-        auto ret = curr();
-        ++d_curr;
-        return ret;
-    }
-
-    auto consume_maybe(token_type tt) -> bool;
-    auto consume_only(token_type tt, std::source_location loc = std::source_location::current()) -> token;
-    auto consume_i64() -> std::int64_t;
-    auto consume_u64() -> std::uint64_t;
-    auto peek(token_type tt) -> bool;
-    auto peek_next(token_type tt) -> bool;
-
-    template <typename Func>
-    auto consume_comma_separated_list(token_type tt, Func&& callback) -> void
-    {
-        if (consume_maybe(tt)) { // Empty list
-            return;
-        }
-        callback(); // Parse first
-        while (!peek(tt)) {
-            consume_only(token_type::comma);
-            callback();
-        }
-        consume_only(tt);
-    }
-};
+auto print_token(token tok) -> void;
+auto to_string(token_type tt) -> std::string_view;
     
 }
